Fixed warning with gcc 4.3.3 which showed that the fromBuffer() methods were not allocating enough memory and always over running the memory allocated

+++ openjaus/branches/OpenJAUSv3.3b/libjaus/src/type/jausLong.c
@@ -56,7 +56,7 @@ JausBoolean jausLongFromBuffer(JausLong *jLong, unsigned char *buf, unsigned int
 #ifdef JAUS_BIG_ENDIAN
 	int i = 0;
 #endif	
-	long tempLong = 0;
+	long long tempLong = 0;
 	
 	if(bufferSizeBytes < JAUS_LONG_SIZE_BYTES)
 	{
